packages/indexer/pyproject.toml
[project]
name = "vibe-defi-indexer"
version = "1.0.0"
description = "DeFi yield data indexer for VibeDeFi"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.29.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
